+++ CHANGELOG.md
@@ -35,6 +35,14 @@
 - You can now specify custom log format and log level in `dakara_base.config.create_logger` with arguments `custom_log_format` and `custom_log_level`.
 - Access to and create user-level stored Dakara config files with `dakara_base.config.get_config_file` and `dakara_base.config.create_config_file`.
 
+### Changed
+
+- `progress_bar`: the progress bar now displays percentage.
+
+### Fixed
+
+- `progress_bar`: When an exception was raised within a progress bar, it would prevent to stop the capture of stderr, leading to hide any further log entry.
+
 ## 1.1.0 - 2019-09-16
 
 ### Added