+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 
 ## Unreleased
 
+## 1.2.0 - 2019-11-10
+
 ### Added
 
 - You can now specify custom log format and log level in `dakara_base.config.create_logger` with arguments `custom_log_format` and `custom_log_level`.